Revert Isotopic Compositions (Abundances) to IUPAC 2009
I believe the method taken with the IUPAC 2013 data is erroneous. As it is based on a single sample measurement (terrestrial source), and thus not representative of the potential interval possible in nature.
Some median or mean likely weighted by the measurements is more appropriate, which I expect IUPAC 2009 Column 9 is (in contrast to column 4 from IUPAC 2013).